+++ CHANGELOG.md
@@ -15,6 +15,7 @@
   [`child_process.spawn(command, args, options)`](https://nodejs.org/api/child_process.html#child_process_child_process_spawn_command_args_options),
   [`options.spawn`](https://github.com/ehmicky/nve/blob/master/README.md#nveversionrange-args-options)
   must now be used instead of `options`.
+- `stdio` now defaults to `pipe` instead of `inherit`
 
 ## Features